i) symlinks removed, ii) adding more subfolder structures inside the system layer directory
