+++ CMakeLists.txt
@@ -1,6 +1,9 @@
 file(GLOB astra_SRC 
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/astra-sim/system/*.cc"
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/astra-sim/workload/*.cc")
+	"${CMAKE_CURRENT_SOURCE_DIR}/src/astra-sim/workload/*.cc"
+	"${CMAKE_CURRENT_SOURCE_DIR}/src/astra-sim/system/collective/*.cc"
+	"${CMAKE_CURRENT_SOURCE_DIR}/src/astra-sim/system/topology/*.cc"
+	"${CMAKE_CURRENT_SOURCE_DIR}/src/astra-sim/system/memory/*.cc")
 file(GLOB astra_test_SRC "${CMAKE_CURRENT_SOURCE_DIR}/test/*.cc")
 
 add_library(AstraSim ${astra_SRC})
