+++ src/astra-sim/system/QueueLevels.hh
@@ -36,7 +36,7 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <sstream>
 #include <assert.h>
 #include "QueueLevelHandler.hh"
-#include "RingTopology.hh"
+#include "src/astra-sim/system/topology/RingTopology.hh"
 
 namespace AstraSim{
     class QueueLevels{
