+++ src/astra-sim/system/Sys.cc
@@ -27,15 +27,15 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include "SimRecvCaller.hh"
 #include "QueueLevels.hh"
 #include "MemBus.hh"
-#include "DoubleBinaryTreeAllReduce.hh"
-#include "LocalRingGlobalBinaryTree.hh"
-#include "LocalRingNodeA2AGlobalDBT.hh"
-#include "Ring.hh"
-#include "AllToAll.hh"
+#include "src/astra-sim/system/collective/DoubleBinaryTreeAllReduce.hh"
+#include "src/astra-sim/system/collective/Ring.hh"
+#include "src/astra-sim/system/collective/AllToAll.hh"
 
-#include "DoubleBinaryTreeTopology.hh"
-#include "Torus3D.hh"
-#include "RingTopology.hh"
+#include "src/astra-sim/system/topology/DoubleBinaryTreeTopology.hh"
+#include "src/astra-sim/system/topology/Torus3D.hh"
+#include "src/astra-sim/system/topology/RingTopology.hh"
+#include "src/astra-sim/system/topology/LocalRingGlobalBinaryTree.hh"
+#include "src/astra-sim/system/topology/LocalRingNodeA2AGlobalDBT.hh"
 
 #include "RendezvousSendData.hh"
 #include "RendezvousRecvData.hh"
