--- src/astra-sim/system/astra-sim
@@ -1,1 +1,0 @@
-../../astra-sim
