+++ src/astra-sim/system/collective/Algorithm.cc
