+++ src/astra-sim/system/collective/Algorithm.hh
@@ -19,8 +19,8 @@ SOFTWARE.
 Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 *******************************************************************************/
 
-#ifndef __COMPLEXLOGICALTOPOLOGY_HH__
-#define __COMPLEXLOGICALTOPOLOGY_HH__
+#ifndef __ALGORITHM_HH__
+#define __ALGORITHM_HH__
 
 #include <map>
 #include <math.h>
@@ -35,16 +35,31 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <chrono>
 #include <sstream>
 #include <assert.h>
-#include "Common.hh"
-#include "LogicalTopology.hh"
+#include "src/astra-sim/system/Common.hh"
+#include "src/astra-sim/system/Callable.hh"
+#include "src/astra-sim/system/BaseStream.hh"
+#include "src/astra-sim/system/CallData.hh"
+#include "src/astra-sim/system/topology/LogicalTopology.hh"
 
 namespace AstraSim{
-    class ComplexLogicalTopology: public LogicalTopology{
+    class Algorithm:public Callable{
     public:
-        ComplexLogicalTopology(){this->complexity=LogicalTopology::Complexity::Complex;}
-        virtual ~ComplexLogicalTopology()=default;
-        virtual int get_num_of_dimensions() override{return 1;}
-        virtual BasicLogicalTopology* get_basic_topology_at_dimension(int dimension,ComType type) override{return NULL;};
+        enum class Name{Ring,DoubleBinaryTree,AllToAll};
+        Name name;
+        int id;
+        BaseStream *stream;
+        LogicalTopology *logicalTopology;
+        int data_size;
+        int final_data_size;
+        ComType comType;
+        bool enabled;
+        int layer_num;
+        Algorithm(int layer_num);
+        virtual ~Algorithm()= default;
+        virtual void run(EventType event,CallData *data)=0;
+        virtual void exit();
+        virtual void init(BaseStream *stream);
+        virtual void call(EventType event,CallData *data);
     };
 }
 #endif
