+++ src/astra-sim/system/memory/SimpleMemory.hh
@@ -1,7 +1,7 @@
 #ifndef __SIMPLEMEMORY_HH__
 #define __SIMPLEMEMORY_HH__
-#include "AstraMemoryAPI.hh"
-#include "AstraNetworkAPI.hh"
+#include "src/astra-sim/system/AstraMemoryAPI.hh"
+#include "src/astra-sim/system/AstraNetworkAPI.hh"
 namespace AstraSim{
 class SimpleMemory: public AstraMemoryAPI{
 public:
