+++ src/astra-sim/system/topology/BasicLogicalTopology.hh
@@ -19,8 +19,8 @@ SOFTWARE.
 Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 *******************************************************************************/
 
-#ifndef __ALGORITHM_HH__
-#define __ALGORITHM_HH__
+#ifndef __BASICLOGICALTOPOLOGY_HH__
+#define __BASICLOGICALTOPOLOGY_HH__
 
 #include <map>
 #include <math.h>
@@ -35,31 +35,19 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <chrono>
 #include <sstream>
 #include <assert.h>
-#include "Common.hh"
-#include "Callable.hh"
-#include "BaseStream.hh"
-#include "CallData.hh"
+#include "src/astra-sim/system/Common.hh"
 #include "LogicalTopology.hh"
 
 namespace AstraSim{
-    class Algorithm:public Callable{
+    class BasicLogicalTopology: public LogicalTopology{
     public:
-        enum class Name{Ring,DoubleBinaryTree,AllToAll};
-        Name name;
-        int id;
-        BaseStream *stream;
-        LogicalTopology *logicalTopology;
-        int data_size;
-        int final_data_size;
-        ComType comType;
-        bool enabled;
-        int layer_num;
-        Algorithm(int layer_num);
-        virtual ~Algorithm()= default;
-        virtual void run(EventType event,CallData *data)=0;
-        virtual void exit();
-        virtual void init(BaseStream *stream);
-        virtual void call(EventType event,CallData *data);
+        enum class BasicTopology{Ring,BinaryTree};
+        BasicTopology basic_topology;
+        BasicLogicalTopology(BasicTopology basic_topology){this->basic_topology=basic_topology;
+            this->complexity=LogicalTopology::Complexity::Basic;}
+        virtual ~BasicLogicalTopology()=default;
+        int get_num_of_dimensions() override{return 1;};
+        BasicLogicalTopology* get_basic_topology_at_dimension(int dimension,ComType type) override{return this;}
     };
 }
 #endif
