+++ src/astra-sim/system/topology/BinaryTree.hh
@@ -19,8 +19,8 @@ SOFTWARE.
 Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 *******************************************************************************/
 
-#ifndef __ALLTOALL_HH__
-#define __ALLTOALL_HH__
+#ifndef __BINARYTREE_HH__
+#define __BINARYTREE_HH__
 
 #include <map>
 #include <math.h>
@@ -35,20 +35,31 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <chrono>
 #include <sstream>
 #include <assert.h>
-#include "Common.hh"
-#include "Ring.hh"
-#include "RingTopology.hh"
-#include "CallData.hh"
+#include "src/astra-sim/system/Common.hh"
+#include "BasicLogicalTopology.hh"
+#include "Node.hh"
 
 namespace AstraSim{
-    class AllToAll:public Ring{
+    class BinaryTree:public BasicLogicalTopology{
     public:
-        AllToAll(ComType type,int id,int layer_num,RingTopology *allToAllTopology,int data_size,
-                 RingTopology::Direction direction,PacketRouting routing,
-                 InjectionPolicy injection_policy,bool boost_mode);
-        void run(EventType event,CallData *data);
-        void process_max_count();
-        int get_non_zero_latency_packets();
+        enum class TreeType{RootMax,RootMin};
+        enum class Type{Leaf,Root,Intermediate};
+        int total_tree_nodes;
+        int start;
+        TreeType tree_type;
+        int stride;
+        Node *tree;
+        virtual ~BinaryTree();
+        std::map<int,Node*> node_list;
+        BinaryTree(int id,TreeType tree_type,int total_tree_nodes,int start,int stride);
+        Node* initialize_tree(int depth,Node* parent);
+        void build_tree(Node* node);
+        int get_parent_id(int id);
+        int get_left_child_id(int id);
+        int get_right_child_id(int id);
+        Type get_node_type(int id);
+        bool is_enabled(int id){if(id%stride==0){return true;}return false;};
+        void print(Node *node);
     };
 }
 #endif
