+++ src/astra-sim/system/topology/DoubleBinaryTreeTopology.hh
@@ -19,8 +19,8 @@ SOFTWARE.
 Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 *******************************************************************************/
 
-#ifndef __LOCALRINGNODEA2AGLOBALDBT_HH__
-#define __LOCALRINGNODEA2AGLOBALDBT_HH__
+#ifndef __DOUBLEBINARYTREETOPOLOGY_HH__
+#define __DOUBLEBINARYTREETOPOLOGY_HH__
 
 #include <map>
 #include <math.h>
@@ -35,16 +35,18 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <chrono>
 #include <sstream>
 #include <assert.h>
-#include "Common.hh"
+#include "src/astra-sim/system/Common.hh"
 #include "ComplexLogicalTopology.hh"
-#include "DoubleBinaryTreeTopology.hh"
-
+#include "LocalRingGlobalBinaryTree.hh"
 namespace AstraSim{
-    class LocalRingNodeA2AGlobalDBT:public ComplexLogicalTopology{
+    class DoubleBinaryTreeTopology:public ComplexLogicalTopology{
     public:
-        DoubleBinaryTreeTopology *global_all_reduce_dimension;
-        LocalRingNodeA2AGlobalDBT(int id,int total_tree_nodes,int start,int stride,int local_dim);
-        ~LocalRingNodeA2AGlobalDBT();
+        int counter;
+        LocalRingGlobalBinaryTree *DBMAX;
+        LocalRingGlobalBinaryTree *DBMIN;
+        LogicalTopology *get_topology();
+        ~DoubleBinaryTreeTopology();
+        DoubleBinaryTreeTopology(int id,int total_tree_nodes,int start,int stride,int local_dim);
     };
 }
 #endif
