+++ src/astra-sim/system/topology/LocalRingGlobalBinaryTree.hh
@@ -19,8 +19,8 @@ SOFTWARE.
 Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 *******************************************************************************/
 
-#ifndef __BINARYTREE_HH__
-#define __BINARYTREE_HH__
+#ifndef __LOCALRINGGLOBALBINARYTREE_HH__
+#define __LOCALRINGGLOBALBINARYTREE_HH__
 
 #include <map>
 #include <math.h>
@@ -35,31 +35,18 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <chrono>
 #include <sstream>
 #include <assert.h>
-#include "Common.hh"
-#include "BasicLogicalTopology.hh"
-#include "Node.hh"
+#include "src/astra-sim/system/Common.hh"
+#include "RingTopology.hh"
+#include "BinaryTree.hh"
+#include "ComplexLogicalTopology.hh"
 
 namespace AstraSim{
-    class BinaryTree:public BasicLogicalTopology{
+    class LocalRingGlobalBinaryTree:public ComplexLogicalTopology{
     public:
-        enum class TreeType{RootMax,RootMin};
-        enum class Type{Leaf,Root,Intermediate};
-        int total_tree_nodes;
-        int start;
-        TreeType tree_type;
-        int stride;
-        Node *tree;
-        virtual ~BinaryTree();
-        std::map<int,Node*> node_list;
-        BinaryTree(int id,TreeType tree_type,int total_tree_nodes,int start,int stride);
-        Node* initialize_tree(int depth,Node* parent);
-        void build_tree(Node* node);
-        int get_parent_id(int id);
-        int get_left_child_id(int id);
-        int get_right_child_id(int id);
-        Type get_node_type(int id);
-        bool is_enabled(int id){if(id%stride==0){return true;}return false;};
-        void print(Node *node);
+        RingTopology *local_dimension;
+        BinaryTree *global_dimension;
+        LocalRingGlobalBinaryTree(int id,BinaryTree::TreeType tree_type,int total_tree_nodes,int start,int stride,int local_dim);
+        ~LocalRingGlobalBinaryTree();
     };
 }
 #endif
