+++ src/astra-sim/system/topology/LogicalTopology.cc
