+++ src/astra-sim/system/topology/Node.cc
