+++ src/astra-sim/system/topology/RingTopology.hh
@@ -19,8 +19,8 @@ SOFTWARE.
 Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 *******************************************************************************/
 
-#ifndef __LOCALRINGGLOBALBINARYTREE_HH__
-#define __LOCALRINGGLOBALBINARYTREE_HH__
+#ifndef __RINGTOPOLOGY_HH__
+#define __RINGTOPOLOGY_HH__
 
 #include <map>
 #include <math.h>
@@ -35,18 +35,30 @@ Author : Saeed Rashidi (saeed.rashidi@gatech.edu)
 #include <chrono>
 #include <sstream>
 #include <assert.h>
-#include "Common.hh"
-#include "RingTopology.hh"
-#include "BinaryTree.hh"
-#include "ComplexLogicalTopology.hh"
+#include "src/astra-sim/system/Common.hh"
+#include "BasicLogicalTopology.hh"
 
 namespace AstraSim{
-    class LocalRingGlobalBinaryTree:public ComplexLogicalTopology{
+    class RingTopology:public BasicLogicalTopology{
     public:
-        RingTopology *local_dimension;
-        BinaryTree *global_dimension;
-        LocalRingGlobalBinaryTree(int id,BinaryTree::TreeType tree_type,int total_tree_nodes,int start,int stride,int local_dim);
-        ~LocalRingGlobalBinaryTree();
+        enum class Direction{Clockwise,Anticlockwise};
+        enum class Dimension{Local,Vertical,Horizontal};
+        std::string name;
+        int id;
+        int next_node_id;
+        int previous_node_id;
+        int offset;
+        int total_nodes_in_ring;
+        int index_in_ring;
+        Dimension dimension;
+        RingTopology(Dimension dimension,int id,int total_nodes_in_ring,int index_in_ring,int offset);
+        void find_neighbors();
+        virtual int get_receiver_node(int node_id,Direction direction);
+        virtual int get_sender_node(int node_id,Direction direction);
+        int get_nodes_in_ring();
+        bool is_enabled();
+    private:
+        std::map<int,int> id_to_index;
     };
 }
 #endif
