--- src/astra-sim/workload/astra-sim
@@ -1,1 +1,0 @@
-../../astra-sim
